# zapzapclone-ui/openapi/back.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/message": {
      "post": {
        "tags": ["Message"],
        "operationId": "postMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Message"],
        "operationId": "patchStatusChange",
        "parameters": [
          {
            "name": "chatId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["SENT", "RECEIVED", "SEEN"]
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/message/upload-media": {
      "post": {
        "tags": ["Message"],
        "operationId": "postMessageWithFile",
        "parameters": [
          {
            "name": "chatId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "get": {
        "tags": ["Chat"],
        "operationId": "getAllChats",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Chat"],
        "operationId": "postChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": ["User"],
        "operationId": "getAllUsersButMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/message/{chatId}": {
      "get": {
        "tags": ["Message"],
        "operationId": "getMessagesByChat",
        "parameters": [
          {
            "name": "chatId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MessageRequest": {
        "required": ["chatId", "content", "receiverId", "type"],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "receiverId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["TEXT", "IMAGE", "AUDIO", "DOCUMENT", "VIDEO"]
          },
          "chatId": {
            "type": "string"
          }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/User"
          },
          "receiver": {
            "$ref": "#/components/schemas/User"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": ["SENT", "RECEIVED", "SEEN"]
          },
          "type": {
            "type": "string",
            "enum": ["TEXT", "IMAGE", "AUDIO", "DOCUMENT", "VIDEO"]
          },
          "senderId": {
            "type": "string"
          },
          "receiverId": {
            "type": "string"
          },
          "chat": {
            "$ref": "#/components/schemas/Chat"
          },
          "mediaFilePath": {
            "type": "string"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": ["SENT", "RECEIVED", "SEEN"]
          },
          "type": {
            "type": "string",
            "enum": ["TEXT", "IMAGE", "AUDIO", "DOCUMENT", "VIDEO"]
          },
          "senderId": {
            "type": "string"
          },
          "receiverId": {
            "type": "string"
          },
          "chat": {
            "$ref": "#/components/schemas/Chat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "file": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "chatsAsSender": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chat"
            }
          },
          "chatsAsReceiver": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chat"
            }
          },
          "online": {
            "type": "boolean"
          }
        }
      },
      "ChatRequest": {
        "required": ["receiverId"],
        "type": "object",
        "properties": {
          "receiverId": {
            "type": "string"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "chatName": {
            "type": "string"
          },
          "unreadMessagesCount": {
            "type": "integer",
            "format": "int64"
          },
          "lastMessage": {
            "type": "string"
          },
          "lastMessageTime": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "receiverId": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          },
          "reciverOnline": {
            "type": "boolean"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "online": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "keycloak": {
        "type": "oauth2",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "password": {
            "authorizationUrl": "http://localhost:6969/realms/zapzap/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:6969/realms/zapzap/protocol/openid-connect/token",
            "scopes": {}
          }
        }
      },
      "bearerAuth": {
        "type": "oauth2",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:6969/realms/zapzap/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:6969/realms/zapzap/protocol/openid-connect/token",
            "scopes": {}
          }
        }
      }
    }
  }
}