# zapzapclone-ui/src/app/api-services/models/chat-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Message } from '../models/message';
import { UserResponse } from '../models/user-response';
export interface ChatResponse {
  chatName?: string;
  id?: string;
  lastMessage?: string;
  lastMessageTime?: string;
  messages?: Array<Message>;
  receiverId?: string;
  reciverOnline?: boolean;
  senderId?: string;
  unreadMessagesCount?: number;
  users?: Array<UserResponse>;
}