# zapzapclone-ui/src/app/api-services/models/message-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Chat } from '../models/chat';
export interface MessageResponse {
  chat?: Chat;
  content?: string;
  createdAt: string;
  file?: string;
  id?: number;
  receiverId?: string;
  senderId?: string;
  state?: 'SENT' | 'RECEIVED' | 'SEEN';
  type?: 'TEXT' | 'IMAGE' | 'AUDIO' | 'DOCUMENT' | 'VIDEO';
  updatedAt?: string;
}